MASSIVE UDPATE: extreme optimization
i tried plugging in my friend's data into my program, which was around 200 mb, which is fucking massive. it took like 10 minutes and it still crashed at the end.
therefore, i optimized the function to unfuckthedata to hell and back. It used to be O(n^2) or even O(n^3), but now it's a solid O(n).
This made it so that the thing that took 10 minutes now took 10 seconds or even less.
That is FUCKING MASSIVE. I had to do some weird ass magic that i dont even fully understand tbh.

ALSO, i thought there was some time-related bug, but nah, im just stupid and used incorect save data